Refactor dashboard component and add testing
