+++ tests/sample.test.js
@@ -1,0 +1,7 @@
+import { describe, it, expect } from 'vitest';
+
+describe('Sample test', () => {
+  it('should run correctly', () => {
+    expect(1 + 1).toBe(2);
+  });
+});
